fix: resolve 5 POS screen bugs in payment and display logic
- Fix receipt print data never set (preparePrintData called as reference, not invoked)
- Fix impossible qtyCart.length < 0 condition blocking Reset/Hold modals
- Fix parseInt truncating decimal amounts in payment validation (use parseFloat)
- Remove debug console.log left in production
- Fix Location field showing user last name instead of warehouse name

+++ resources/pos/src/frontend/components/cart-product/paymentModel/CashPaymentModel.js
@@ -345,8 +345,8 @@ const CashPaymentModel = (props) => {
                     onClick={(event) => {
                         if (cashPaymentValue.received_amount !== undefined) {
                             if (
-                                parseInt(cashPaymentValue.received_amount) <
-                                parseInt(grandTotal)
+                                parseFloat(cashPaymentValue.received_amount) <
+                                parseFloat(grandTotal)
                             ) {
                                 dispatch(
                                     addToast({
@@ -372,8 +372,8 @@ const CashPaymentModel = (props) => {
                     onClick={(event) => {
                         if (cashPaymentValue.received_amount !== undefined) {
                             if (
-                                parseInt(cashPaymentValue.received_amount) <
-                                parseInt(grandTotal)
+                                parseFloat(cashPaymentValue.received_amount) <
+                                parseFloat(grandTotal)
                             ) {
                                 dispatch(
                                     addToast({
